trustistics/backend/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
pymongo>=4.7.0
python-dotenv>=1.0.1
qrcode[pil]>=7.4.2
web3>=6.19.0
pydantic>=2.7.0
python-multipart>=0.0.9
